Moved TripList to trips folder
